requirements.txt
amqp==2.3.2
billiard==3.5.0.4
blinker==1.4
celery==4.2.0
click==6.7
Flask==1.0.2
Flask-Mail==0.9.1
imutils==0.4.6
itsdangerous==0.24
Jinja2==2.10
kombu==4.2.1
MarkupSafe==1.0
numpy==1.14.5
opencv-python==3.4.1.15
Pillow==5.1.0
pkg-resources==0.0.0
pyocr==0.5.1
PyPDF2==1.26.0
pytz==2018.5
six==1.11.0
vine==1.1.4
Wand==0.4.4
Werkzeug==0.14.1
XlsxWriter==1.0.5
